Fix for issue 1933 - combine relationship config with other config.
