+++ src/EntityFramework.Relational.Design/ReverseEngineering/Configuration/NavigationConfiguration.cs
@@ -1,0 +1,31 @@
+﻿// Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using JetBrains.Annotations;
+using Microsoft.Data.Entity.Metadata;
+using Microsoft.Data.Entity.Utilities;
+
+namespace Microsoft.Data.Entity.Relational.Design.ReverseEngineering.Configuration
+{
+    public class NavigationConfiguration
+    {
+        public NavigationConfiguration([NotNull]EntityConfiguration entityConfiguration,
+            [NotNull]IForeignKey foreignKey, [NotNull]string dependentEndNavigationPropertyName,
+            [CanBeNull]string principalEndNavigationPropertyName)
+        {
+            Check.NotNull(entityConfiguration, nameof(entityConfiguration));
+            Check.NotNull(foreignKey, nameof(foreignKey));
+            Check.NotEmpty(dependentEndNavigationPropertyName, nameof(dependentEndNavigationPropertyName));
+
+            EntityConfiguration = entityConfiguration;
+            ForeignKey = foreignKey;
+            DependentEndNavigationPropertyName = dependentEndNavigationPropertyName;
+            PrincipalEndNavigationPropertyName = principalEndNavigationPropertyName;
+        }
+
+        public virtual EntityConfiguration EntityConfiguration { get;[param: NotNull]private set; }
+        public virtual IForeignKey ForeignKey { get;[param: NotNull]private set; }
+        public virtual string DependentEndNavigationPropertyName { get;[param: NotNull]private set; }
+        public virtual string PrincipalEndNavigationPropertyName { get;[param: NotNull]private set; }
+    }
+}
